+++ app/helpers/tabs/names/descriptions_helper.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+# html used in tabsets
+module Tabs
+  module Names
+    module DescriptionsHelper
+    end
+  end
+end
